update visualization for sdv.evaluation
