+++ sdv/evaluation/__init__.py
@@ -4,9 +4,17 @@ from sdv.evaluation.evaluation import (
     evaluate_quality,
     run_diagnostic,
 )
+from sdv.evaluation._visualization import (
+    get_cardinality_plot,
+    get_column_pair_plot,
+    get_column_plot,
+)
 
 
 __all__ = (
     'evaluate_quality',
     'run_diagnostic',
+    'get_cardinality_plot',
+    'get_column_pair_plot',
+    'get_column_plot',
 )
