+++ tests/unit/evaluation/test__visualization.py
@@ -1,23 +1,60 @@
 import re
-from unittest.mock import patch
+from unittest.mock import call, patch
 
+import numpy as np
 import pandas as pd
 import pytest
 
 from sdv.errors import VisualizationUnavailableError
-from sdv.evaluation.single_table import (
+from sdv.evaluation._visualization import (
+    _get_column_pair_plot,
+    _get_column_plot,
+    _prepare_data_visualization,
+    get_cardinality_plot,
     get_column_pair_plot,
     get_column_plot,
 )
+from sdv.metadata._single_table import _SingleTableMetadata
 from sdv.metadata.metadata import Metadata
 
 
-@patch('sdmetrics.visualization.get_column_plot')
-@patch('sdv.evaluation.single_table._prepare_data_visualization')
-def test_get_column_plot_continuous_data(mock_prepare, mock_get_plot):
-    """Test the ``get_column_plot`` with continuous data.
+def test__prepare_data_visualization():
+    """Test ``_prepare_data_visualization``."""
+    # Setup
+    np.random.seed(0)
+    metadata = _SingleTableMetadata.load_from_dict({
+        'columns': {
+            'col1': {'sdtype': 'datetime', 'datetime_format': '%Y-%m-%d'},
+            'col2': {'sdtype': 'numerical'},
+        }
+    })
+    column_names = ['col1', 'col2']
+    sample_size = 2
+    data = pd.DataFrame({
+        'col1': ['2021-01-01', '2021-02-01', '2021-03-01'],
+        'col2': [4, 5, 6],
+    })
 
-    Test that when we call ``get_column_plot`` with continuous data (datetime or numerical)
+    # Run
+    result = _prepare_data_visualization(data, metadata, column_names, sample_size)
+
+    # Assert
+    expected_result = pd.DataFrame(
+        {
+            'col1': pd.to_datetime(['2021-03-01', '2021-02-01']),
+            'col2': [6, 5],
+        },
+        index=[2, 1],
+    )
+    pd.testing.assert_frame_equal(result, expected_result)
+
+
+@patch('sdmetrics.visualization.get_column_plot')
+@patch('sdv.evaluation._visualization._prepare_data_visualization')
+def test__get_column_plot_continuous_data(mock_prepare, mock_get_plot):
+    """Test the ``_get_column_plot`` with continuous data.
+
+    Test that when we call ``_get_column_plot`` with continuous data (datetime or numerical)
     this will choose to use the ``distplot`` as ``plot_type``.
     """
     # Setup
@@ -29,7 +66,7 @@ def test_get_column_plot_continuous_data(mock_prepare, mock_get_plot):
     mock_prepare.side_effect = [data1, data2]
 
     # Run
-    plot = get_column_plot(data1, data2, metadata, 'col')
+    plot = _get_column_plot(data1, data2, metadata, 'col')
 
     # Assert
     mock_get_plot.assert_called_once_with(data1, data2, 'col', plot_type='distplot')
@@ -37,11 +74,11 @@ def test_get_column_plot_continuous_data(mock_prepare, mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-@patch('sdv.evaluation.single_table._prepare_data_visualization')
-def test_get_column_plot_continuous_data_metadata(mock_prepare, mock_get_plot):
-    """Test the ``get_column_plot`` with continuous data.
+@patch('sdv.evaluation._visualization._prepare_data_visualization')
+def test__get_column_plot_continuous_data_metadata(mock_prepare, mock_get_plot):
+    """Test the ``_get_column_plot`` with continuous data.
 
-    Test that when we call ``get_column_plot`` with continuous data (datetime or numerical)
+    Test that when we call ``_get_column_plot`` with continuous data (datetime or numerical)
     this will choose to use the ``distplot`` as ``plot_type``. Uses Metadata.
     """
     # Setup
@@ -52,7 +89,7 @@ def test_get_column_plot_continuous_data_metadata(mock_prepare, mock_get_plot):
     mock_prepare.side_effect = [data1, data2]
 
     # Run
-    plot = get_column_plot(data1, data2, metadata, 'col')
+    plot = _get_column_plot(data1, data2, metadata, 'col')
 
     # Assert
     mock_get_plot.assert_called_once_with(data1, data2, 'col', plot_type='distplot')
@@ -60,11 +97,11 @@ def test_get_column_plot_continuous_data_metadata(mock_prepare, mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-@patch('sdv.evaluation.single_table._prepare_data_visualization')
-def test_get_column_plot_discrete_data(mock_prepare, mock_get_plot):
-    """Test the ``get_column_plot`` with discrete data.
+@patch('sdv.evaluation._visualization._prepare_data_visualization')
+def test__get_column_plot_discrete_data(mock_prepare, mock_get_plot):
+    """Test the ``_get_column_plot`` with discrete data.
 
-    Test that when we call ``get_column_plot`` with discrete data (categorical or boolean)
+    Test that when we call ``_get_column_plot`` with discrete data (categorical or boolean)
     this will choose to use the ``bar`` as ``plot_type``.
     """
     # Setup
@@ -76,7 +113,7 @@ def test_get_column_plot_discrete_data(mock_prepare, mock_get_plot):
     mock_prepare.side_effect = [data1, data2]
 
     # Run
-    plot = get_column_plot(data1, data2, metadata, 'col')
+    plot = _get_column_plot(data1, data2, metadata, 'col')
 
     # Assert
     mock_get_plot.assert_called_once_with(data1, data2, 'col', plot_type='bar')
@@ -84,11 +121,11 @@ def test_get_column_plot_discrete_data(mock_prepare, mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-@patch('sdv.evaluation.single_table._prepare_data_visualization')
-def test_get_column_plot_discrete_data_metadata(mock_prepare, mock_get_plot):
-    """Test the ``get_column_plot`` with discrete data.
+@patch('sdv.evaluation._visualization._prepare_data_visualization')
+def test__get_column_plot_discrete_data_metadata(mock_prepare, mock_get_plot):
+    """Test the ``_get_column_plot`` with discrete data.
 
-    Test that when we call ``get_column_plot`` with discrete data (categorical or boolean)
+    Test that when we call ``_get_column_plot`` with discrete data (categorical or boolean)
     this will choose to use the ``bar`` as ``plot_type``. Uses Metadata.
     """
     # Setup
@@ -99,7 +136,7 @@ def test_get_column_plot_discrete_data_metadata(mock_prepare, mock_get_plot):
     mock_prepare.side_effect = [data1, data2]
 
     # Run
-    plot = get_column_plot(data1, data2, metadata, 'col')
+    plot = _get_column_plot(data1, data2, metadata, 'col')
 
     # Assert
     mock_get_plot.assert_called_once_with(data1, data2, 'col', plot_type='bar')
@@ -107,11 +144,11 @@ def test_get_column_plot_discrete_data_metadata(mock_prepare, mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-@patch('sdv.evaluation.single_table._prepare_data_visualization')
-def test_get_column_plot_discrete_data_with_distplot(mock_prepare, mock_get_plot):
-    """Test the ``get_column_plot`` with discrete data.
+@patch('sdv.evaluation._visualization._prepare_data_visualization')
+def test__get_column_plot_discrete_data_with_distplot(mock_prepare, mock_get_plot):
+    """Test the ``_get_column_plot`` with discrete data.
 
-    Test that when we call ``get_column_plot`` with discrete data (categorical or boolean)
+    Test that when we call ``_get_column_plot`` with discrete data (categorical or boolean)
     and pass in the ``distplot`` it will call the ``sdmetrics.visualization.get_column_plot``
     with it and not switch to ``bar``.
     """
@@ -124,7 +161,7 @@ def test_get_column_plot_discrete_data_with_distplot(mock_prepare, mock_get_plot
     mock_prepare.side_effect = [data1, data2]
 
     # Run
-    plot = get_column_plot(data1, data2, metadata, 'col', plot_type='distplot')
+    plot = _get_column_plot(data1, data2, metadata, 'col', plot_type='distplot')
 
     # Assert
     mock_get_plot.assert_called_once_with(data1, data2, 'col', plot_type='distplot')
@@ -132,11 +169,11 @@ def test_get_column_plot_discrete_data_with_distplot(mock_prepare, mock_get_plot
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-@patch('sdv.evaluation.single_table._prepare_data_visualization')
-def test_get_column_plot_discrete_data_with_distplot_metadata(mock_prepare, mock_get_plot):
-    """Test the ``get_column_plot`` with discrete data.
+@patch('sdv.evaluation._visualization._prepare_data_visualization')
+def test__get_column_plot_discrete_data_with_distplot_metadata(mock_prepare, mock_get_plot):
+    """Test the ``_get_column_plot`` with discrete data.
 
-    Test that when we call ``get_column_plot`` with discrete data (categorical or boolean)
+    Test that when we call ``_get_column_plot`` with discrete data (categorical or boolean)
     and pass in the ``distplot`` it will call the ``sdmetrics.visualization.get_column_plot``
     with it and not switch to ``bar``. Uses Metadata.
     """
@@ -148,7 +185,7 @@ def test_get_column_plot_discrete_data_with_distplot_metadata(mock_prepare, mock
     mock_prepare.side_effect = [data1, data2]
 
     # Run
-    plot = get_column_plot(data1, data2, metadata, 'col', plot_type='distplot')
+    plot = _get_column_plot(data1, data2, metadata, 'col', plot_type='distplot')
 
     # Assert
     mock_get_plot.assert_called_once_with(data1, data2, 'col', plot_type='distplot')
@@ -156,11 +193,11 @@ def test_get_column_plot_discrete_data_with_distplot_metadata(mock_prepare, mock
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-@patch('sdv.evaluation.single_table._prepare_data_visualization')
-def test_get_column_plot_invalid_sdtype(mock_prepare, mock_get_plot):
-    """Test the ``get_column_plot`` with sdtype that can't be plotted.
+@patch('sdv.evaluation._visualization._prepare_data_visualization')
+def test__get_column_plot_invalid_sdtype(mock_prepare, mock_get_plot):
+    """Test the ``_get_column_plot`` with sdtype that can't be plotted.
 
-    Test that when we call ``get_column_plot`` with an sdtype that can't be plotted, this raises
+    Test that when we call ``_get_column_plot`` with an sdtype that can't be plotted, this raises
     an error.
     """
     # Setup
@@ -177,15 +214,15 @@ def test_get_column_plot_invalid_sdtype(mock_prepare, mock_get_plot):
         "supported visualization. To visualize this data anyways, please add a 'plot_type'."
     )
     with pytest.raises(VisualizationUnavailableError, match=error_msg):
-        get_column_plot(data1, data2, metadata, 'col')
+        _get_column_plot(data1, data2, metadata, 'col')
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-@patch('sdv.evaluation.single_table._prepare_data_visualization')
-def test_get_column_plot_invalid_sdtype_metadata(mock_prepare, mock_get_plot):
-    """Test the ``get_column_plot`` with sdtype that can't be plotted.
+@patch('sdv.evaluation._visualization._prepare_data_visualization')
+def test__get_column_plot_invalid_sdtype_metadata(mock_prepare, mock_get_plot):
+    """Test the ``_get_column_plot`` with sdtype that can't be plotted.
 
-    Test that when we call ``get_column_plot`` with an sdtype that can't be plotted, this raises
+    Test that when we call ``_get_column_plot`` with an sdtype that can't be plotted, this raises
     an error. Uses Metadata.
     """
     # Setup
@@ -201,15 +238,15 @@ def test_get_column_plot_invalid_sdtype_metadata(mock_prepare, mock_get_plot):
         "supported visualization. To visualize this data anyways, please add a 'plot_type'."
     )
     with pytest.raises(VisualizationUnavailableError, match=error_msg):
-        get_column_plot(data1, data2, metadata, 'col')
+        _get_column_plot(data1, data2, metadata, 'col')
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-@patch('sdv.evaluation.single_table._prepare_data_visualization')
-def test_get_column_plot_invalid_sdtype_with_plot_type(mock_prepare, mock_get_plot):
-    """Test the ``get_column_plot`` with sdtype that can't be plotted.
+@patch('sdv.evaluation._visualization._prepare_data_visualization')
+def test__get_column_plot_invalid_sdtype_with_plot_type(mock_prepare, mock_get_plot):
+    """Test the ``_get_column_plot`` with sdtype that can't be plotted.
 
-    Test that when we call ``get_column_plot`` with an sdtype that can't be plotted, but passing
+    Test that when we call ``_get_column_plot`` with an sdtype that can't be plotted, but passing
     ``plot_type`` it will attempt to plot it using the ``sdmetrics.visualization.get_column_plot``.
     """
     # Setup
@@ -221,7 +258,7 @@ def test_get_column_plot_invalid_sdtype_with_plot_type(mock_prepare, mock_get_pl
     mock_prepare.side_effect = [data1, data2]
 
     # Run
-    plot = get_column_plot(data1, data2, metadata, 'col', plot_type='bar')
+    plot = _get_column_plot(data1, data2, metadata, 'col', plot_type='bar')
 
     # Assert
     mock_get_plot.assert_called_once_with(data1, data2, 'col', plot_type='bar')
@@ -229,11 +266,11 @@ def test_get_column_plot_invalid_sdtype_with_plot_type(mock_prepare, mock_get_pl
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-@patch('sdv.evaluation.single_table._prepare_data_visualization')
-def test_get_column_plot_invalid_sdtype_with_plot_type_metadata(mock_prepare, mock_get_plot):
-    """Test the ``get_column_plot`` with sdtype that can't be plotted.
+@patch('sdv.evaluation._visualization._prepare_data_visualization')
+def test__get_column_plot_invalid_sdtype_with_plot_type_metadata(mock_prepare, mock_get_plot):
+    """Test the ``_get_column_plot`` with sdtype that can't be plotted.
 
-    Test that when we call ``get_column_plot`` with an sdtype that can't be plotted, but passing
+    Test that when we call ``_get_column_plot`` with an sdtype that can't be plotted, but passing
     ``plot_type`` it will attempt to plot it using the ``sdmetrics.visualization.get_column_plot``.
     """
     # Setup
@@ -244,7 +281,7 @@ def test_get_column_plot_invalid_sdtype_with_plot_type_metadata(mock_prepare, mo
     mock_prepare.side_effect = [data1, data2]
 
     # Run
-    plot = get_column_plot(data1, data2, metadata, 'col', plot_type='bar')
+    plot = _get_column_plot(data1, data2, metadata, 'col', plot_type='bar')
 
     # Assert
     mock_get_plot.assert_called_once_with(data1, data2, 'col', plot_type='bar')
@@ -252,8 +289,8 @@ def test_get_column_plot_invalid_sdtype_with_plot_type_metadata(mock_prepare, mo
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-def test_get_column_plot_real_data_none(mock_get_plot):
-    """Test ``get_column_plot`` when ``real_data`` is None."""
+def test__get_column_plot_real_data_none(mock_get_plot):
+    """Test ``_get_column_plot`` when ``real_data`` is None."""
     # Setup
     data = pd.DataFrame({'col': [1, 2, 3]})
     metadata = Metadata()
@@ -261,7 +298,7 @@ def test_get_column_plot_real_data_none(mock_get_plot):
     metadata.add_column('col', 'table', sdtype='numerical')
 
     # Run
-    plot = get_column_plot(None, data, metadata, 'col')
+    plot = _get_column_plot(None, data, metadata, 'col')
 
     # Assert
     mock_get_plot.call_args[0][1].equals(data)
@@ -270,8 +307,8 @@ def test_get_column_plot_real_data_none(mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-def test_get_column_plot_synthetic_data_none(mock_get_plot):
-    """Test ``get_column_plot`` when ``synthetic_data`` is None."""
+def test__get_column_plot_synthetic_data_none(mock_get_plot):
+    """Test ``_get_column_plot`` when ``synthetic_data`` is None."""
     # Setup
     data = pd.DataFrame({'col': [1, 2, 3]})
     metadata = Metadata()
@@ -279,7 +316,7 @@ def test_get_column_plot_synthetic_data_none(mock_get_plot):
     metadata.add_column('col', 'table', sdtype='numerical')
 
     # Run
-    plot = get_column_plot(data, None, metadata, 'col')
+    plot = _get_column_plot(data, None, metadata, 'col')
 
     # Assert
     mock_get_plot.call_args[0][0].equals(data)
@@ -288,10 +325,10 @@ def test_get_column_plot_synthetic_data_none(mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_plot')
-def test_get_column_plot_with_datetime_sdtype(mock_get_plot):
-    """Test the ``get_column_plot`` with datetime sdtype.
+def test__get_column_plot_with_datetime_sdtype(mock_get_plot):
+    """Test the ``_get_column_plot`` with datetime sdtype.
 
-    Test that when we call ``get_column_plot`` with ``datetime`` this will parse it using the
+    Test that when we call ``_get_column_plot`` with ``datetime`` this will parse it using the
     datetime format provided in the metadata and it will cast it to ``datetime64``.
     """
     # Setup
@@ -302,7 +339,7 @@ def test_get_column_plot_with_datetime_sdtype(mock_get_plot):
     metadata.add_column('datetime', 'table', sdtype='datetime', datetime_format='%Y-%m-%d')
 
     # Run
-    plot = get_column_plot(real_data, synthetic_data, metadata, 'datetime')
+    plot = _get_column_plot(real_data, synthetic_data, metadata, 'datetime')
 
     # Assert
     expected_real_data = pd.DataFrame({'datetime': pd.to_datetime(['2021-02-01', '2021-12-01'])})
@@ -316,10 +353,10 @@ def test_get_column_plot_with_datetime_sdtype(mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_pair_plot')
-def test_get_column_pair_plot_with_continous_data(mock_get_plot):
-    """Test ``get_column_pair_plot`` with continuous data.
+def test__get_column_pair_plot_with_continous_data(mock_get_plot):
+    """Test ``_get_column_pair_plot`` with continuous data.
 
-    Test that when we call ``get_column_pair_plot`` with ``continuous`` data, this will
+    Test that when we call ``_get_column_pair_plot`` with ``continuous`` data, this will
     automatically choose to use the ``scatter`` plot instead of the ``heatmap``.
     """
     # Setup
@@ -338,7 +375,7 @@ def test_get_column_pair_plot_with_continous_data(mock_get_plot):
     metadata.add_column('date', 'table', sdtype='datetime')
 
     # Run
-    plot = get_column_pair_plot(real_data, synthetic_data, metadata, columns)
+    plot = _get_column_pair_plot(real_data, synthetic_data, metadata, columns)
 
     # Assert
     expected_real_data = pd.DataFrame({
@@ -357,11 +394,11 @@ def test_get_column_pair_plot_with_continous_data(mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_pair_plot')
-@patch('sdv.evaluation.single_table._prepare_data_visualization')
-def test_get_column_pair_plot_with_discrete_data(mock_prepare, mock_get_plot):
-    """Test the ``get_column_pair_plot`` when using discrete data.
+@patch('sdv.evaluation._visualization._prepare_data_visualization')
+def test__get_column_pair_plot_with_discrete_data(mock_prepare, mock_get_plot):
+    """Test the ``_get_column_pair_plot`` when using discrete data.
 
-    Test that the ``get_column_pair_plot`` will automatically use ``heatmap`` if the data
+    Test that the ``_get_column_pair_plot`` will automatically use ``heatmap`` if the data
     provided is discrete.
     """
     # Setup
@@ -375,7 +412,7 @@ def test_get_column_pair_plot_with_discrete_data(mock_prepare, mock_get_plot):
     mock_prepare.side_effect = [real_data, synthetic_data]
 
     # Run
-    plot = get_column_pair_plot(real_data, synthetic_data, metadata, columns)
+    plot = _get_column_pair_plot(real_data, synthetic_data, metadata, columns)
 
     # Assert
     pd.testing.assert_frame_equal(mock_get_plot.call_args[0][0], real_data)
@@ -386,8 +423,8 @@ def test_get_column_pair_plot_with_discrete_data(mock_prepare, mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_pair_plot')
-def test_get_column_pair_plot_with_mixed_data(mock_get_plot):
-    """Test the ``get_column_pair_plot`` with mixed data types.
+def test__get_column_pair_plot_with_mixed_data(mock_get_plot):
+    """Test the ``_get_column_pair_plot`` with mixed data types.
 
     Test that when using both discrete and continuous data, we will be using automatically the
     ``box`` plot.
@@ -402,7 +439,7 @@ def test_get_column_pair_plot_with_mixed_data(mock_get_plot):
     metadata.add_column('counts', 'table', sdtype='numerical')
 
     # Run
-    plot = get_column_pair_plot(real_data, synthetic_data, metadata, columns)
+    plot = _get_column_pair_plot(real_data, synthetic_data, metadata, columns)
 
     # Assert
     pd.testing.assert_frame_equal(mock_get_plot.call_args[0][0], real_data)
@@ -413,8 +450,8 @@ def test_get_column_pair_plot_with_mixed_data(mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_pair_plot')
-def test_get_column_pair_plot_with_forced_plot_type(mock_get_plot):
-    """Test the ``get_column_pair_plot`` with continuous data and fixed plot type.
+def test__get_column_pair_plot_with_forced_plot_type(mock_get_plot):
+    """Test the ``_get_column_pair_plot`` with continuous data and fixed plot type.
 
     Test that when using ``continuous`` data but asking to plot as ``heatmap`` this will still
     force the ``sdmetrics.visualization.get_column_pair_plot`` to use this.
@@ -435,7 +472,7 @@ def test_get_column_pair_plot_with_forced_plot_type(mock_get_plot):
     metadata.add_column('date', 'table', sdtype='datetime')
 
     # Run
-    plot = get_column_pair_plot(real_data, synthetic_data, metadata, columns, plot_type='heatmap')
+    plot = _get_column_pair_plot(real_data, synthetic_data, metadata, columns, plot_type='heatmap')
 
     # Assert
     expected_real_data = pd.DataFrame({
@@ -455,10 +492,10 @@ def test_get_column_pair_plot_with_forced_plot_type(mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_pair_plot')
-def test_get_column_pair_plot_with_invalid_sdtype(mock_get_plot):
-    """Test the ``get_column_pair_plot`` with sdtype that can't be plotted.
+def test__get_column_pair_plot_with_invalid_sdtype(mock_get_plot):
+    """Test the ``_get_column_pair_plot`` with sdtype that can't be plotted.
 
-    Test that when we call ``get_column_pair_plot`` with an sdtype that can't be plotted,
+    Test that when we call ``_get_column_pair_plot`` with an sdtype that can't be plotted,
     this raises an error.
     """
     # Setup
@@ -482,12 +519,12 @@ def test_get_column_pair_plot_with_invalid_sdtype(mock_get_plot):
         "supported visualization. To visualize this data anyways, please add a 'plot_type'."
     )
     with pytest.raises(VisualizationUnavailableError, match=error_msg):
-        get_column_pair_plot(real_data, synthetic_data, metadata, columns)
+        _get_column_pair_plot(real_data, synthetic_data, metadata, columns)
 
 
 @patch('sdmetrics.visualization.get_column_pair_plot')
-def test_get_column_pair_plot_with_invalid_sdtype_and_plot_type(mock_get_plot):
-    """Test the ``get_column_pair_plot`` with sdtype that can't be plotted but providing plot type.
+def test__get_column_pair_plot_with_invalid_sdtype_and_plot_type(mock_get_plot):
+    """Test the ``_get_column_pair_plot`` with sdtype that can't be plotted but providing plot type.
 
     Test that when providing the ``plot_type`` for an sdtype that can't be plotted, this will be
     plotted.
@@ -508,7 +545,7 @@ def test_get_column_pair_plot_with_invalid_sdtype_and_plot_type(mock_get_plot):
     metadata.add_column('id', 'table', sdtype='id')
 
     # Run
-    plot = get_column_pair_plot(real_data, synthetic_data, metadata, columns, plot_type='heatmap')
+    plot = _get_column_pair_plot(real_data, synthetic_data, metadata, columns, plot_type='heatmap')
 
     # Assert
     pd.testing.assert_frame_equal(mock_get_plot.call_args[0][0], real_data)
@@ -519,8 +556,8 @@ def test_get_column_pair_plot_with_invalid_sdtype_and_plot_type(mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_pair_plot')
-def test_get_column_pair_plot_with_sample_size(mock_get_plot):
-    """Test ``get_column_pair_plot`` with ``sample_size`` parameter."""
+def test__get_column_pair_plot_with_sample_size(mock_get_plot):
+    """Test ``_get_column_pair_plot`` with ``sample_size`` parameter."""
     # Setup
     columns = ['amount', 'price']
     real_data = pd.DataFrame({
@@ -537,7 +574,7 @@ def test_get_column_pair_plot_with_sample_size(mock_get_plot):
     metadata.add_column('price', 'table', sdtype='numerical')
 
     # Run
-    get_column_pair_plot(real_data, synthetic_data, metadata, columns, sample_size=2)
+    _get_column_pair_plot(real_data, synthetic_data, metadata, columns, sample_size=2)
 
     # Assert
     real_subsample = mock_get_plot.call_args[0][0]
@@ -549,8 +586,8 @@ def test_get_column_pair_plot_with_sample_size(mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_pair_plot')
-def test_get_column_pair_plot_with_sample_size_metadata(mock_get_plot):
-    """Test ``get_column_pair_plot`` with ``sample_size`` parameter with Metadata."""
+def test__get_column_pair_plot_with_sample_size_metadata(mock_get_plot):
+    """Test ``_get_column_pair_plot`` with ``sample_size`` parameter with Metadata."""
     # Setup
     columns = ['amount', 'price']
     real_data = pd.DataFrame({
@@ -570,7 +607,7 @@ def test_get_column_pair_plot_with_sample_size_metadata(mock_get_plot):
     metadata = Metadata.load_from_dict(metadata_dict)
 
     # Run
-    get_column_pair_plot(real_data, synthetic_data, metadata, columns, sample_size=2)
+    _get_column_pair_plot(real_data, synthetic_data, metadata, columns, sample_size=2)
 
     # Assert
     real_subsample = mock_get_plot.call_args[0][0]
@@ -582,8 +619,8 @@ def test_get_column_pair_plot_with_sample_size_metadata(mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_pair_plot')
-def test_get_column_pair_plot_with_sample_size_too_big(mock_get_plot):
-    """Test ``get_column_pair_plot`` when ``sample_size`` is bigger than the length of the data."""
+def test__get_column_pair_plot_with_sample_size_too_big(mock_get_plot):
+    """Test ``_get_column_pair_plot`` when ``sample_size`` is bigger than the length of the data."""
     # Setup
     columns = ['amount', 'price']
     real_data = pd.DataFrame({
@@ -600,7 +637,7 @@ def test_get_column_pair_plot_with_sample_size_too_big(mock_get_plot):
     metadata.add_column('price', 'table', sdtype='numerical')
 
     # Run
-    plot = get_column_pair_plot(real_data, synthetic_data, metadata, columns, sample_size=10)
+    plot = _get_column_pair_plot(real_data, synthetic_data, metadata, columns, sample_size=10)
 
     # Assert
     pd.testing.assert_frame_equal(mock_get_plot.call_args[0][0], real_data)
@@ -611,8 +648,8 @@ def test_get_column_pair_plot_with_sample_size_too_big(mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_pair_plot')
-def test__get_column_pair_plot_with_real_data_none(mock_get_plot):
-    """Test ``get_column_pair_plot`` when ``real_data`` is None."""
+def test___get_column_pair_plot_with_real_data_none(mock_get_plot):
+    """Test ``_get_column_pair_plot`` when ``real_data`` is None."""
     # Setup
     columns = ['amount', 'price']
     real_data = None
@@ -626,7 +663,7 @@ def test__get_column_pair_plot_with_real_data_none(mock_get_plot):
     metadata.add_column('price', 'table', sdtype='numerical')
 
     # Run
-    plot = get_column_pair_plot(real_data, synthetic_data, metadata, columns)
+    plot = _get_column_pair_plot(real_data, synthetic_data, metadata, columns)
 
     # Assert
     assert mock_get_plot.call_args[0][0] is None
@@ -637,8 +674,8 @@ def test__get_column_pair_plot_with_real_data_none(mock_get_plot):
 
 
 @patch('sdmetrics.visualization.get_column_pair_plot')
-def test__get_column_pair_plot_with_synthetic_data_none(mock_get_plot):
-    """Test ``get_column_pair_plot`` when ``synthetic_data`` is None."""
+def test___get_column_pair_plot_with_synthetic_data_none(mock_get_plot):
+    """Test ``_get_column_pair_plot`` when ``synthetic_data`` is None."""
     # Setup
     columns = ['amount', 'price']
     real_data = pd.DataFrame({
@@ -652,7 +689,7 @@ def test__get_column_pair_plot_with_synthetic_data_none(mock_get_plot):
     metadata.add_column('price', 'table', sdtype='numerical')
 
     # Run
-    plot = get_column_pair_plot(real_data, synthetic_data, metadata, columns)
+    plot = _get_column_pair_plot(real_data, synthetic_data, metadata, columns)
 
     # Assert
     pd.testing.assert_frame_equal(mock_get_plot.call_args[0][0], real_data)
@@ -660,3 +697,201 @@ def test__get_column_pair_plot_with_synthetic_data_none(mock_get_plot):
     assert mock_get_plot.call_args[0][2] == columns
     assert mock_get_plot.call_args[0][3] == 'scatter'
     assert plot == mock_get_plot.return_value
+
+
+@patch('sdv.evaluation._visualization._get_column_plot')
+def test_get_column_plot(mock_plot):
+    """Test the ``get_column_plot``.
+
+    Ensure that the ``get_column_plot`` is being called with the ``_SingleTableMetadata`` object
+    and the expected table.
+    """
+    # Setup
+    table1 = pd.DataFrame({'col': [1, 2, 3]})
+    table2 = pd.DataFrame({'col': [2, 1, 3]})
+    data1 = {'table': table1}
+    data2 = {'table': table2}
+    metadata = Metadata()
+    metadata.detect_table_from_dataframe('table', table1)
+    mock_plot.return_value = 'plot'
+
+    # Run
+    plot = get_column_plot(data1, data2, metadata, 'table', 'col')
+
+    # Assert
+    call_metadata = metadata.tables['table']
+    mock_plot.assert_called_once_with(table1, table2, call_metadata, 'col', None)
+    assert plot == 'plot'
+
+
+@patch('sdv.evaluation._visualization._get_column_plot')
+def test_get_column_plot_only_real_or_synthetic(mock_plot):
+    """Test that ``get_column_plot`` works when only real or synthetic data is provided."""
+    # Setup
+    table1 = pd.DataFrame({'col': [1, 2, 3]})
+    data1 = {'table': table1}
+    metadata = Metadata()
+    metadata.detect_table_from_dataframe('table', table1)
+    mock_plot.return_value = 'plot'
+
+    # Run
+    get_column_plot(data1, None, metadata, 'table', 'col')
+    get_column_plot(None, data1, metadata, 'table', 'col')
+
+    # Assert
+    call_metadata = metadata.tables['table']
+    mock_plot.assert_has_calls([
+        ((table1, None, call_metadata, 'col', None), {}),
+        ((None, table1, call_metadata, 'col', None), {}),
+    ])
+
+
+@patch('sdv.evaluation._visualization._get_column_pair_plot')
+def test_get_column_pair_plot(mock_plot):
+    """Test that ``get_column_pair_plot`` is being called with the expected objects."""
+    # Setup
+    table1 = pd.DataFrame({'col1': [1, 2, 3], 'col2': [3, 2, 1]})
+    table2 = pd.DataFrame({'col1': [2, 1, 3], 'col2': [1, 2, 3]})
+    data1 = {'table': table1}
+    data2 = {'table': table2}
+    metadata = Metadata()
+    metadata.detect_table_from_dataframe('table', table1)
+    mock_plot.return_value = 'plot'
+
+    # Run
+    plot = get_column_pair_plot(data1, data2, metadata, 'table', ['col1', 'col2'], sample_size=2)
+
+    # Assert
+    call_metadata = metadata.tables['table']
+    mock_plot.assert_called_once_with(
+        real_data=table1,
+        synthetic_data=table2,
+        metadata=call_metadata,
+        column_names=['col1', 'col2'],
+        plot_type=None,
+        sample_size=2,
+    )
+    assert plot == 'plot'
+
+
+@patch('sdv.evaluation._visualization._get_column_pair_plot')
+def test_get_column_pair_plot_only_real_or_synthetic(mock_plot):
+    """Test that ``get_column_pair_plot`` works when only real or synthetic data is provided."""
+    # Setup
+    table1 = pd.DataFrame({'col1': [1, 2, 3], 'col2': [3, 2, 1]})
+    data1 = {'table': table1}
+    metadata = Metadata()
+    metadata.detect_table_from_dataframe('table', table1)
+    mock_plot.return_value = 'plot'
+
+    # Run
+    get_column_pair_plot(data1, None, metadata, 'table', ['col1', 'col2'], sample_size=2)
+    get_column_pair_plot(None, data1, metadata, 'table', ['col1', 'col2'], sample_size=2)
+
+    # Assert
+    call_metadata = metadata.tables['table']
+    call1 = call(
+        real_data=table1,
+        synthetic_data=None,
+        metadata=call_metadata,
+        column_names=['col1', 'col2'],
+        plot_type=None,
+        sample_size=2,
+    )
+    call2 = call(
+        real_data=None,
+        synthetic_data=table1,
+        metadata=call_metadata,
+        column_names=['col1', 'col2'],
+        plot_type=None,
+        sample_size=2,
+    )
+    mock_plot.assert_has_calls([call1, call2])
+
+
+@patch('sdmetrics.visualization.get_cardinality_plot')
+def test_get_cardinality_plot(mock_plot):
+    """Test it calls ``get_column_cardinality_plot`` in sdmetrics with the parent primary key."""
+    # Setup
+    data1 = {
+        'table1': pd.DataFrame({'col1': [1, 2, 3], 'col2': [3, 2, 1]}),
+        'table2': pd.DataFrame({'col1': [2, 2, 3], 'col2': [6, 7, 8]}),
+    }
+    data2 = {
+        'table1': pd.DataFrame({'col1': [2, 1, 3], 'col2': [1, 2, 3]}),
+        'table2': pd.DataFrame({'col1': [2, 2, 3], 'col2': [6, 7, 8]}),
+    }
+    metadata_dict = {
+        'tables': {
+            'table1': {
+                'columns': {'col1': {'sdtype': 'numerical'}, 'col2': {'sdtype': 'numerical'}},
+                'primary_key': 'col1',
+            },
+            'table2': {
+                'columns': {'col1': {'sdtype': 'numerical'}, 'col2': {'sdtype': 'numerical'}}
+            },
+        },
+        'relationships': [
+            {
+                'parent_table_name': 'table1',
+                'child_table_name': 'table2',
+                'parent_primary_key': 'col1',
+                'child_foreign_key': 'col1',
+            }
+        ],
+        'METADATA_SPEC_VERSION': 'MULTI_TABLE_V1',
+    }
+    metadata = Metadata.load_from_dict(metadata_dict)
+    mock_plot.return_value = 'plot'
+
+    # Run
+    plot = get_cardinality_plot(data1, data2, 'table2', 'table1', 'col1', metadata)
+
+    # Assert
+    mock_plot.assert_called_once_with(data1, data2, 'table2', 'table1', 'col1', 'col1', 'bar')
+    assert plot == 'plot'
+
+
+@patch('sdmetrics.visualization.get_cardinality_plot')
+def test_get_cardinality_plot_plot_type(mock_plot):
+    """Test it calls ``get_column_cardinality_plot`` with different ``plot_type``."""
+    # Setup
+    data1 = {
+        'table1': pd.DataFrame({'col1': [1, 2, 3], 'col2': [3, 2, 1]}),
+        'table2': pd.DataFrame({'col1': [2, 2, 3], 'col2': [6, 7, 8]}),
+    }
+    data2 = {
+        'table1': pd.DataFrame({'col1': [2, 1, 3], 'col2': [1, 2, 3]}),
+        'table2': pd.DataFrame({'col1': [2, 2, 3], 'col2': [6, 7, 8]}),
+    }
+    metadata_dict = {
+        'tables': {
+            'table1': {
+                'columns': {'col1': {'sdtype': 'numerical'}, 'col2': {'sdtype': 'numerical'}},
+                'primary_key': 'col1',
+            },
+            'table2': {
+                'columns': {'col1': {'sdtype': 'numerical'}, 'col2': {'sdtype': 'numerical'}}
+            },
+        },
+        'relationships': [
+            {
+                'parent_table_name': 'table1',
+                'child_table_name': 'table2',
+                'parent_primary_key': 'col1',
+                'child_foreign_key': 'col1',
+            }
+        ],
+        'METADATA_SPEC_VERSION': 'MULTI_TABLE_V1',
+    }
+    metadata = Metadata.load_from_dict(metadata_dict)
+    mock_plot.return_value = 'plot'
+
+    # Run
+    plot = get_cardinality_plot(
+        data1, data2, 'table2', 'table1', 'col1', metadata, plot_type='distplot'
+    )
+
+    # Assert
+    mock_plot.assert_called_once_with(data1, data2, 'table2', 'table1', 'col1', 'col1', 'distplot')
+    assert plot == 'plot'
